+++ plugins/outputs/warp10/warp10.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
+	"math"
 	"net/http"
 	"sort"
 	"strconv"
@@ -23,34 +24,41 @@ const (
 
 // Warp10 output plugin
 type Warp10 struct {
-	Prefix             string
-	WarpURL            string
-	Token              string
+	Prefix             string            `toml:"prefix"`
+	WarpURL            string            `toml:"warp_url"`
+	Token              string            `toml:"token"`
 	Timeout            internal.Duration `toml:"timeout"`
-	PrintErrorBody     bool
-	MaxStringErrorSize int
+	PrintErrorBody     bool              `toml:"print_error_body"`
+	MaxStringErrorSize int               `toml:"max_string_error_size"`
 	client             *http.Client
 	tls.ClientConfig
 }
 
 var sampleConfig = `
-  # prefix for metrics class Name
+  # Prefix to add to the measurement.
   prefix = "telegraf."
-  ## POST HTTP(or HTTPS) ##
-  # Url name of the Warp 10 server
+
+  # URL of the Warp 10 server
   warp_url = "http://localhost:8080"
-  # Token to access your app on warp 10
+
+  # Write token to access your app on warp 10
   token = "Token"
-  # Warp 10 query timeout, by default 15s
-  timeout = "15s"
-  ## Optional Print Warp 10 error body
+
+  # Warp 10 query timeout
+  # timeout = "15s"
+
+  ## Print Warp 10 error body
   # print_error_body = false
-  ## Optional Max string error Size
+
+  ## Max string error size
   # max_string_error_size = 511
+
   ## Optional TLS Config
   # tls_ca = "/etc/telegraf/ca.pem"
   # tls_cert = "/etc/telegraf/cert.pem"
   # tls_key = "/etc/telegraf/key.pem"
+  ## Use TLS but skip chain & host verification
+  # insecure_skip_verify = false
 `
 
 // MetricLine Warp 10 metrics
@@ -94,7 +102,7 @@ func (w *Warp10) Connect() error {
 }
 
 // GenWarp10Payload compute Warp 10 metrics payload
-func (w *Warp10) GenWarp10Payload(metrics []telegraf.Metric, now time.Time) string {
+func (w *Warp10) GenWarp10Payload(metrics []telegraf.Metric) string {
 	collectString := make([]string, 0)
 	for _, mm := range metrics {
 
@@ -102,7 +110,7 @@ func (w *Warp10) GenWarp10Payload(metrics []telegraf.Metric, now time.Time) stri
 
 			metric := &MetricLine{
 				Metric:    fmt.Sprintf("%s%s", w.Prefix, mm.Name()+"."+field.Key),
-				Timestamp: now.UnixNano() / 1000,
+				Timestamp: mm.Time().UnixNano() / 1000,
 			}
 
 			metricValue, err := buildValue(field.Value)
@@ -125,10 +133,7 @@ func (w *Warp10) GenWarp10Payload(metrics []telegraf.Metric, now time.Time) stri
 
 // Write metrics to Warp10
 func (w *Warp10) Write(metrics []telegraf.Metric) error {
-
-	var now = time.Now()
-	payload := w.GenWarp10Payload(metrics, now)
-
+	payload := w.GenWarp10Payload(metrics)
 	if payload == "" {
 		return nil
 	}
@@ -177,17 +182,21 @@ func buildValue(v interface{}) (string, error) {
 	var retv string
 	switch p := v.(type) {
 	case int64:
-		retv = intToString(int64(p))
+		retv = intToString(p)
 	case string:
 		retv = fmt.Sprintf("'%s'", strings.Replace(p, "'", "\\'", -1))
 	case bool:
-		retv = boolToString(bool(p))
+		retv = boolToString(p)
 	case uint64:
-		retv = uIntToString(uint64(p))
+		if p <= uint64(math.MaxInt64) {
+			retv = strconv.FormatInt(int64(p), 10)
+		} else {
+			retv = strconv.FormatInt(math.MaxInt64, 10)
+		}
 	case float64:
 		retv = floatToString(float64(p))
 	default:
-		retv = "'" + strings.Replace(fmt.Sprintf("%s", p), "'", "\\'", -1) + "'"
+		return "", fmt.Errorf("unsupported type: %T", v)
 	}
 	return retv, nil
 }
@@ -215,7 +224,7 @@ func (w *Warp10) SampleConfig() string {
 
 // Description get description
 func (w *Warp10) Description() string {
-	return "Configuration for Warp server to send metrics to"
+	return "Write metrics to Warp 10"
 }
 
 // Close close